requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-tinymce==3.3.0
psycopg2==2.9.1
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
